Current

+++ system/images/03.serviceImages/dns/home/init.sh
@@ -6,7 +6,10 @@ chown -R bind /var/run/named
 mkdir -p /var/log/named
 chown -R bind /var/log/named
 
-rm /var/run/named/named.pid
+	if test -f /var/run/named/named.pid
+		then
+			rm /var/run/named/named.pid
+	fi
 /usr/sbin/named -c /etc/bind/named.conf -u bind 
 
 mkdir -p /engines/var/run/
